# Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 30-Day Coding Challenge API
  description: |
    API for a progressive 30-day coding challenge platform supporting 5 languages.
    
    ## Features
    - 30 sequential daily lessons per language
    - Automated code grading
    - Timezone-aware rollback punishment
    - Admin tooling and audit logging
    
    ## Languages Supported
    - JavaScript (Node.js)
    - Python
    - Java
    - Go
    - C#
  version: 1.0.0
  contact:
    email: support@example.com

servers:
  - url: http://localhost:3000/api
    description: Development server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Lessons
    description: Lesson content endpoints
  - name: Submissions
    description: Code submission endpoints
  - name: Progress
    description: User progress endpoints
  - name: Admin
    description: Admin-only endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        timezone:
          type: string
        progress:
          type: array
          items:
            $ref: '#/components/schemas/Progress'
        createdAt:
          type: string
          format: date-time

    Progress:
      type: object
      properties:
        language:
          type: string
          enum: [javascript, python, java, go, csharp]
        currentDay:
          type: integer
          minimum: 1
          maximum: 30
        lastPassedDay:
          type: integer
          minimum: 0
          maximum: 30
        failedDay:
          type: integer
          nullable: true
        failedAt:
          type: string
          format: date-time
          nullable: true
        attemptCount:
          type: integer
        completedAt:
          type: string
          format: date-time
          nullable: true

    Lesson:
      type: object
      properties:
        _id:
          type: string
        language:
          type: string
        day:
          type: integer
        title:
          type: string
        objectives:
          type: array
          items:
            type: string
        contentHtml:
          type: string
        examples:
          type: array
          items:
            type: object
        exercise:
          type: object
        tests:
          type: array
          items:
            type: object
        difficulty:
          type: integer
        estimatedMinutes:
          type: integer

    Submission:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        language:
          type: string
        day:
          type: integer
        code:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, error]
        resultDetails:
          type: object
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                timezone:
                  type: string
                  default: UTC
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Validation error
        '409':
          description: Email already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials

  /auth/profile:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile

  /lessons/{language}:
    get:
      tags: [Lessons]
      summary: Get all lessons for a language
      parameters:
        - name: language
          in: path
          required: true
          schema:
            type: string
            enum: [javascript, python, java, go, csharp]
      responses:
        '200':
          description: List of lessons

  /lessons/{language}/{day}:
    get:
      tags: [Lessons]
      summary: Get specific lesson
      parameters:
        - name: language
          in: path
          required: true
          schema:
            type: string
        - name: day
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lesson details
        '404':
          description: Lesson not found

  /submissions:
    post:
      tags: [Submissions]
      summary: Submit code for grading
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [language, day, code]
              properties:
                language:
                  type: string
                day:
                  type: integer
                code:
                  type: string
      responses:
        '201':
          description: Submission created

  /submissions/{id}:
    get:
      tags: [Submissions]
      summary: Get submission by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Submission details

  /progress:
    get:
      tags: [Progress]
      summary: Get all progress
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Progress for all languages

  /progress/{language}:
    get:
      tags: [Progress]
      summary: Get progress for a language
      security:
        - bearerAuth: []
      parameters:
        - name: language
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Progress details

  /admin/progress/override:
    post:
      tags: [Admin]
      summary: Override user progress
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, language, currentDay, reason]
              properties:
                userId:
                  type: string
                language:
                  type: string
                currentDay:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Progress overridden